% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeCOVbin}
\alias{makeCOVbin}
\title{make the bins with equal number of observations
or using user-defined breaks}
\usage{
makeCOVbin(COV.data, N.covbin = NULL, breaks.data = NULL)
}
\arguments{
\item{COV.data}{numeric vector that need to make bins}

\item{N.covbin}{the number of bins}

\item{breaks.data}{user-defined breaks}
}
\value{
information of the binning with summary
}
\description{
make the bins with equal number of observations
or using user-defined breaks
}
\examples{
makeCOVbin(orig.data$TIME,7)
}
\author{
Eun-Kyung Lee \email{lee.eunk@gmail.com}
}

