\name{origdata}
\alias{origdata}
\docType{data}
\title{
sample original data
}
\description{
sample original data
}
\usage{data("origdata")}
\format{
  A data frame with 132 observations on the following 5 variables.
  \describe{
    \item{\code{X.ID}}{a numeric vector}
    \item{\code{AMT}}{a numeric vector}
    \item{\code{TIME}}{a numeric vector}
    \item{\code{DV}}{a numeric vector}
    \item{\code{WT}}{a numeric vector}
  }
}
\examples{
data(origdata)
## maybe str(orig.data) ; plot(orig.data) ...
}
\keyword{datasets}
