% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_Simdata}
\alias{read_Simdata}
\title{calculate percentiles of original data using bin-related weight
percentiles of simulated data with corresponding confidence interval}
\usage{
read_Simdata(sim.file.name, data.n, sim.n, name.DV)
}
\arguments{
\item{sim.file.name}{file name of simulation, generated from NONMEM with 'NOAPPEND ONEHEAD' options in TABLE statement}

\item{data.n}{number of observations in the original data}

\item{sim.n}{number of simulation}

\item{name.DV}{name of dependent variable in simulated data file}
}
\value{
data.n * sim.n matrix with simulated data
}
\description{
calculate percentiles of original data using bin-related weight
percentiles of simulated data with corresponding confidence interval
}
\examples{
read_Simdata("sample.SIM",data.n=144,sim.n=50,name.DV="DV")
}
\author{
Eun-Kyung Lee \email{lee.eunk@gmail.com}
}
\references{
new paper...
}
\seealso{
\code{\link{asVPC.distanceW}}
}

