\name{simdata}
\alias{simdata}
\docType{data}
\title{
sample sim data
}
\description{
sample sim data
}
\usage{data("simdata")}
\format{
  The format is:
 num [1:132, 1:100] 0 2.82 3.14 5.06 4.8 ...
}
\examples{
data(simdata)
## maybe str(sim.data) ; plot(sim.data) ...
}
\keyword{datasets}
