% Generated by roxygen2 (4.0.2): do not edit by hand
\name{VPC.graph}
\alias{VPC.graph}
\title{calculate percentiles of original data using bin-related weight
percentiles of simulated data with corresponding confidence interval}
\usage{
VPC.graph(orig.data, sim.data, N.timebin, N.sim, q.list = c(0.05, 0.5, 0.95),
  alpha = 0.05, X.name = "TIME", Y.name = "DV", opt.DV.point = FALSE,
  opt.DV.quantile.line = TRUE, opt.SIM.quantile.line = FALSE,
  opt.SIM.quantile.CI.area = TRUE, Y.min = NULL, Y.max = NULL)
}
\arguments{
\item{orig.data}{NONMEM data}

\item{sim.data}{simulated data from NONMEM}

\item{N.timebin}{number of time bin}

\item{N.sim}{number of simulation}

\item{q.list}{list of quantiles for VPC plot}

\item{alpha}{significance level of CI for each quantile}

\item{X.name}{x label in VPC plot}

\item{Y.name}{y label in VPC plot}

\item{opt.DV.point}{option for drawing data points}

\item{opt.SIM.quantile.line}{option for drawing quantiles of simulated data}

\item{opt.SIM.quantile.CI.area}{opeions for drawing confidence area of quantiles for simulated data}

\item{Y.min}{minimum of y axis in VPC plot}

\item{Y.max}{}

\item{Y.max}{maximum of y axis in VPC plot}
}
\description{
calculate percentiles of original data using bin-related weight
percentiles of simulated data with corresponding confidence interval
}
\examples{
VPC.graph(orig.data,sim.data,10,50,Y.max=15)
}
\author{
Eun-Kyung Lee \email{lee.eunk@gmail.com}
}
\references{
new paper...
}

