% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_Simdata.R
\name{read_Simdata}
\alias{read_Simdata}
\title{read the simulated data file generated from NONMEM}
\usage{
read_Simdata(sim.file.name, data.n, sim.n, name.DV)
}
\arguments{
\item{sim.file.name}{file name of simulation, generated from NONMEM with 'NOAPPEND ONEHEAD' options in TABLE statement}

\item{data.n}{number of observations in the original data}

\item{sim.n}{number of simulation}

\item{name.DV}{name of dependent variable in simulated data file}
}
\value{
data.n * sim.n matrix with simulated data
}
\description{
read the simulated data file generated from NONMEM
}
\author{
Eun-Kyung Lee \email{lee.eunk@gmail.com}
}
\references{
new paper...
}
\seealso{
\code{\link{asVPC.distanceW}}
}

